Break up readout.go file.
* Split out prometheus functions in to separate file.
